refactor: extract executor.rs and cli_parsing.rs from pool.rs
Decompose pool.rs (2904 -> 2379 lines) by extracting:
- executor.rs (458 lines): task execution, streaming, system prompt building
- cli_parsing.rs (124 lines): ID generation, permission detection, failure extraction

All public Pool API methods stay in pool.rs. Internal fields made pub(crate).

Closes #192
